Split CallRouter functions
